Added Listing Page and Details Tab and linked it to Main Pagethrough URL parameters. Reorganized some of the files
