+++ package.json
@@ -11,6 +11,7 @@
     "google-map-react": "^1.1.5",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0"
   },
   "scripts": {
